+++ src/main/java/com/dp_ua/iksparser/bot/command/impl/CommandCompetitions.java
@@ -31,7 +31,7 @@ public class CommandCompetitions extends BaseCommand {
     @Override
     protected void perform(Message message) {
         String chatId = message.getChatId();
-        int commandArgument = getCommandArgument(message.getMessageText());
+        long commandArgument = getCommandArgument(message.getMessageText());
         try {
             competitionFacade.showCompetitions(chatId, commandArgument, message.getEditMessageId());
         } catch (ParsingException e) {
